MULE-9765: move current sockets module to the transport compatibility layer
