+++ modules/http/src/main/java/org/mule/runtime/module/http/internal/listener/DefaultHttpListenerConfig.java
@@ -27,9 +27,9 @@ import org.mule.runtime.module.http.api.listener.HttpListenerConfig;
 import org.mule.runtime.module.http.internal.HttpParser;
 import org.mule.runtime.module.http.internal.listener.async.RequestHandler;
 import org.mule.runtime.module.http.internal.listener.matcher.ListenerRequestMatcher;
-import org.mule.runtime.module.socket.internal.DefaultTcpServerSocketProperties;
+import org.mule.runtime.transport.socket.internal.DefaultTcpServerSocketProperties;
 import org.mule.runtime.api.tls.TlsContextFactory;
-import org.mule.runtime.module.socket.api.TcpServerSocketProperties;
+import org.mule.runtime.transport.socket.api.TcpServerSocketProperties;
 import org.mule.runtime.core.util.NetworkUtils;
 import org.mule.runtime.core.util.Preconditions;
 import org.mule.runtime.core.util.StringUtils;
