+++ modules/http/src/main/java/org/mule/runtime/module/http/internal/listener/HttpListenerConnectionManager.java
@@ -16,9 +16,9 @@ import org.mule.runtime.core.api.lifecycle.Initialisable;
 import org.mule.runtime.core.api.lifecycle.InitialisationException;
 import org.mule.runtime.core.config.i18n.CoreMessages;
 import org.mule.runtime.module.http.internal.listener.grizzly.GrizzlyServerManager;
-import org.mule.runtime.module.socket.internal.DefaultTcpServerSocketProperties;
+import org.mule.runtime.transport.socket.internal.DefaultTcpServerSocketProperties;
 import org.mule.runtime.api.tls.TlsContextFactory;
-import org.mule.runtime.module.socket.api.TcpServerSocketProperties;
+import org.mule.runtime.transport.socket.api.TcpServerSocketProperties;
 import org.mule.runtime.core.util.concurrent.ThreadNameHelper;
 
 import com.google.common.collect.Iterables;
