+++ modules/http/src/main/java/org/mule/runtime/module/http/internal/listener/grizzly/GrizzlyServerManager.java
@@ -20,7 +20,7 @@ import org.mule.runtime.module.http.internal.listener.HttpServerManager;
 import org.mule.runtime.module.http.internal.listener.Server;
 import org.mule.runtime.module.http.internal.listener.ServerAddress;
 import org.mule.runtime.api.tls.TlsContextFactory;
-import org.mule.runtime.module.socket.api.TcpServerSocketProperties;
+import org.mule.runtime.transport.socket.api.TcpServerSocketProperties;
 import org.mule.runtime.core.util.concurrent.NamedThreadFactory;
 
 import java.io.IOException;
