+++ modules/http/src/main/java/org/mule/runtime/module/http/internal/request/HttpClientConfiguration.java
@@ -8,7 +8,7 @@ package org.mule.runtime.module.http.internal.request;
 
 import org.mule.runtime.api.tls.TlsContextFactory;
 import org.mule.runtime.module.http.api.requester.proxy.ProxyConfig;
-import org.mule.runtime.module.socket.api.TcpClientSocketProperties;
+import org.mule.runtime.transport.socket.api.TcpClientSocketProperties;
 
 
 public class HttpClientConfiguration
