+++ modules/http/src/main/java/org/mule/runtime/module/http/internal/request/grizzly/SocketConfigTransportCustomizer.java
@@ -6,7 +6,7 @@
  */
 package org.mule.runtime.module.http.internal.request.grizzly;
 
-import org.mule.runtime.module.socket.api.TcpClientSocketProperties;
+import org.mule.runtime.transport.socket.api.TcpClientSocketProperties;
 
 import com.ning.http.client.providers.grizzly.TransportCustomizer;
