--- modules/sockets/src/main/resources/META-INF/mule-module.properties
@@ -1,7 +1,0 @@
-module.name=sockets
-
-artifact.export.classPackages=org.mule.runtime.module.socket.api,\
-                              org.mule.runtime.module.socket.internal,\
-                              org.mule.runtime.module.socket.internal.config
-
-artifact.export.resourcePackages=/META-INF
