--- modules/sockets/src/main/resources/META-INF/spring.handlers
@@ -1,1 +1,0 @@
-http\://www.mulesoft.org/schema/mule/sockets=org.mule.runtime.module.socket.internal.config.SocketsNamespaceHandler
