--- modules/sockets/src/main/resources/META-INF/spring.schemas
@@ -1,1 +1,0 @@
-http\://www.mulesoft.org/schema/mule/sockets/current/mule-sockets.xsd=META-INF/mule-sockets.xsd
