+++ tests/http/src/test/resources/http-request-outbound-attachments-config.xml
@@ -3,11 +3,11 @@
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xmlns:http="http://www.mulesoft.org/schema/mule/http"
       xmlns:tls="http://www.mulesoft.org/schema/mule/tls"
-      xmlns:sockets="http://www.mulesoft.org/schema/mule/sockets"
+      xmlns:sockets="http://www.mulesoft.org/schema/mule/transport/sockets"
       xsi:schemaLocation="
                http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
                http://www.mulesoft.org/schema/mule/tls http://www.mulesoft.org/schema/mule/tls/current/mule-tls.xsd
-               http://www.mulesoft.org/schema/mule/sockets http://www.mulesoft.org/schema/mule/sockets/current/mule-sockets.xsd
+               http://www.mulesoft.org/schema/mule/transport/sockets http://www.mulesoft.org/schema/mule/transport/sockets/current/mule-transport-sockets.xsd
                http://www.mulesoft.org/schema/mule/http http://www.mulesoft.org/schema/mule/http/current/mule-http.xsd">
 
     <http:request-config name="requestConfig" host="localhost" port="${httpPort}" />
