+++ transports/sockets/pom.xml
@@ -2,23 +2,29 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.mule.modules</groupId>
-        <artifactId>mule-modules</artifactId>
+        <groupId>org.mule.transports</groupId>
+        <artifactId>mule-transports</artifactId>
         <version>4.0-SNAPSHOT</version>
     </parent>
-    <artifactId>mule-module-tls</artifactId>
+    <artifactId>mule-transport-sockets</artifactId>
     <packaging>jar</packaging>
-    <name>TLS support module</name>
-    <description>A Mule module for connectivity using TLS/SSL authentication.</description>
+    <name>Sockets configuration module</name>
+    <description>A Mule module for configuring sockets.</description>
 
     <dependencies>
         <dependency>
-            <groupId>org.mule.transports</groupId>
-            <artifactId>mule-transport-sockets</artifactId>
+            <groupId>org.mule.modules</groupId>
+            <artifactId>mule-module-spring-config</artifactId>
             <version>${project.version}</version>
         </dependency>
 
-        <!-- test dependencies -->
+        <!-- Unit tests -->
+        <dependency>
+            <groupId>org.mule.tests</groupId>
+            <artifactId>mule-tests-functional</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
         <dependency>
             <groupId>org.mule</groupId>
             <artifactId>mule-core-tests</artifactId>
@@ -27,8 +33,8 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.mule.tests</groupId>
-            <artifactId>mule-tests-functional</artifactId>
+            <groupId>org.mule.modules</groupId>
+            <artifactId>mule-module-client</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
