+++ transports/sockets/src/main/resources/META-INF/mule-module.properties
@@ -1,0 +1,7 @@
+module.name=sockets
+
+artifact.export.classPackages=org.mule.runtime.transport.socket.api,\
+                              org.mule.runtime.transport.socket.internal,\
+                              org.mule.runtime.transport.socket.internal.config
+
+artifact.export.resourcePackages=/META-INF
