+++ transports/sockets/src/main/resources/META-INF/spring.handlers
@@ -1,0 +1,1 @@
+http\://www.mulesoft.org/schema/mule/transport/sockets=org.mule.runtime.transport.socket.internal.config.SocketsNamespaceHandler
