+++ transports/sockets/src/main/resources/META-INF/spring.schemas
@@ -1,0 +1,1 @@
+http\://www.mulesoft.org/schema/mule/transport/sockets/current/mule-transport-sockets.xsd=META-INF/mule-transport-sockets.xsd
